requirements.txt
fastapi
dotenv
google-generativeai 